restructured test dir
