+++ package.json
@@ -81,9 +81,8 @@
   },
   "ava": {
     "files": [
-      "build/module/test/*.spec.js",
-      "build/module/test/Integration/*.spec.js",
-      "build/module/test/VLIW/*.spec.js"
+      "build/module/test/**/*.spec.js",
+      "build/module/test/**/*.spec.js"
     ]
   }
 }
