+++ src/test/functional/VLIW/doubleloop.spec.ts
@@ -1,8 +1,8 @@
 import test from 'ava';
-import { VLIW } from '../../core/VLIW/VLIW';
-import { VLIWCode } from '../../core/VLIW/VLIWCode';
-import { Code } from '../../core/Common/Code';
-import { VLIWError } from '../../core/VLIW/VLIWError';
+import { VLIW } from '../../../core/VLIW/VLIW';
+import { VLIWCode } from '../../../core/VLIW/VLIWCode';
+import { Code } from '../../../core/Common/Code';
+import { VLIWError } from '../../../core/VLIW/VLIWError';
 
 let vliw: VLIW;
 let code: VLIWCode;
@@ -15,38 +15,41 @@ test.beforeEach('Setup machine', () => {
     superescalarCode = new Code();
 });
 
-test('Bucle.pla is executed properly', t => {
+test('Bucledoble.pla is executed properly', t => {
     const inputVLIW = 
-    `15
+    `10
     2	0 0 0 0	2 0 1 0
     3	1 0 0 0	4 0 1 0	3 4 0 0
     1	5 4 0 0
     0
-    0
-    0
-    1	6 2 0 0
     1	8 0 0 0
+    1	6 2 0 0
     0
     0
-    1	7 4 1 0
     0
-    0
-    1	10 5 0 0 2 1 2
-    1	9 0 1 0`;
+    3	9 0 0 0	7 4 0 0	10 5 0 0 2 1 2`;
 
-    const inputSuperescalar = `11
+    const inputSuperescalar = `18
 	ADDI	R2 R0 #50
 	ADDI	R3 R0 #70
 	ADDI	R4 R0 #40
 	LF	F0 (R4)
-	ADDI	R5 R2 #16
+	ADDI	R5 R2 #5
 LOOP:
 	LF 	F1 (R2)
 	ADDF	F1 F1 F0
-	SF	F1 (R3)
+	SF		F1 (R3)
 	ADDI 	R2 R2 #1
 	ADDI	R3 R3 #1
-	BNE	R2 R5 LOOP`;
+	BNE	R2 R5 LOOP
+	ADDI	R3 R0 #70
+	ADDI	R5 R3 #5
+LOOP2:
+	LF		F1 (R3)
+	MULTF	F1 F1 F0
+	SF		F1 (R3)
+	ADDI	R3 R3 #1
+	BNE	R3 R5 LOOP2`;
 
     superescalarCode.load(inputSuperescalar);
     code.load(inputVLIW, superescalarCode);
@@ -54,7 +57,5 @@ LOOP:
     
     while (vliw.tic() !== VLIWError.ENDEXE) { }
 
-    t.deepEqual(vliw.status.cycle, 212, 'Bad pc at finish');
-
-    
+    t.deepEqual(vliw.status.cycle, 56, 'Doubleloop: Bad pc at finish');
 });
