+++ src/test/unit/core/Collections/Queue.spec.ts
@@ -1,5 +1,5 @@
 import test from 'ava';
-import { Queue } from '../core/Collections/Queue';
+import { Queue } from '../../../../core/Collections/Queue';
 
 test('Queue size behaces as expected', (t) => {
    let queue = new Queue<number>(3);
